Allow test failures to not fail the gradle build
This is useful for CI and the technique is employed in gradle testing where the user can define `ignoreFailures = true` which allows the build process to complete even through there are failures.
Added ignoreFailures to the convention and have it set through to the the task.
Any exceptions thrown by the cucumber runner are caught, if we want to ignore failures then the exception is ignored, otherwise we rethrow the exception