Replaced code that does not work in Gradle 5.x.
The call to "sourceSet.output.classesDir" is deprecated in Gradle 4.x and is removed from the api in Gradle 5.x. It is replaced by sourceSet.output.classesDirs and iterating over each classesDir.

Also the version of the Gradle wrapper is upgraded from 2.1 to 3.5 (unfortunately version 4.x or higher did not work).